fix `UndefVarError` of `@sprintf` in Issue #55
For Julia `v1.0`, it is necessary to load `Printf` module before calling `@sprintf`.(#55 ) So there might be two ways to resolve this issue. One is to add dependence of `Printf` which might bring a new trouble for old versions. Another way is directly to use `error` function with substrings, as in this pull request.